Refactor IR code data handling and connect UI events to presenter
- Change `IrCode` data field type from `String` to `Long` for more efficient signal processing.
- Update `SamsungPattern` and `IRCommand` to use numeric literals instead of string-based hex values.
- Refactor `RemoteControllerUI` to accept lambda callbacks for all remote control actions.
- Connect `MainActivity` to the `RemoteControllerPresenter` by passing command triggers into the UI layer.
- Add debug logging in `RemoteControllerPresenter` to track transmitted IR commands.

+++ app/src/main/java/com/reringuy/remotecontroller/data/pattern/SamsungPattern.kt
@@ -10,7 +10,7 @@ fun samsungPatternConverter(irCode: IrCode): IntArray {
     for (i in 0 until irCode.bits) {
         pattern += 560
 
-        val bit = (irCode.data.toLong() shr i) and 1
+        val bit = (irCode.data shr i) and 1
 
         pattern += if (bit == 1L) 1600 else 560
     }
